# Speed up line rendering with WebGL2 instancing