Resolves https://github.com/processing/p5.js/issues/6091

### Changes

Currently, drawing strokes in WebGL mode is fairly slow. Since each segment/cap/join of a stroke is made of at least 2 triangles, we had to duplicate data per vertex of each triangle on the CPU.

Now that WebGL2 is enabled by default, we can use instanced rendering, which lets us repeat attributes per instance being rendered. This lets the GPU do the copying for us.

Here's a stress test example: https://editor.p5js.org/davepagurek/sketches/cWr5RmNAK

If you toggle between the 1.6.0 CDN `<script>` tag to the uploaded p5.min.js, **the frame rate goes up from ~40fps to ~60fps** on Chrome my M1 Macbook Pro (closer to ~18fps and ~40fps in Firefox.)

If you toggle `version: 2` to `version: 1`, it goes back to copying the vertex data on the CPU and the frame rate drops to ~35fps. This is slower than before since the format required for instancing is a bit less efficient when done on the CPU, but WebGL2 support is pretty broad now so this should only happen on legacy code.

### Future work
The new bottleneck in line rendering seems to be libtess. There are a few things we might consider doing:
- Update how `curveDetail` works to maybe be relative to curve length (or curvature) to save vertices
- Move away from libtess, maybe to [mapbox/earcut](https://github.com/mapbox/earcut) which apparently is faster, but would require extra work to tesselate shapes that aren't on the XY plane

#### PR Checklist

- [x] `npm run lint` passes
- [x] [Inline documentation] is included / updated
- [ ] [Unit tests] are included / updated