@aferriss 2023-05-30
For the stress test would it be clearer to create the lines without using the noise function? I worry that function is also computationally pricey and could be obscuring your results.