@aferriss 2023-05-30
On the demo you linked, I'm actually seeing better performance from the CDN p5, using both webGL1 and webGL2. Your version was giving around 15fps, the CDN was close to 60 when drawing the different shapes.

I also got an error when trying to use your version with webGL1, not sure if maybe that script got built before all of your changes were made, but just want to flag that.