@davepagurek 2023-05-30
> On the demo you linked, I'm actually seeing better performance from the CDN p5, using both webGL1 and webGL2. Your version was giving around 15fps, the CDN was close to 60 when drawing the different shapes.

I've noticed sometimes changes to the html file don't update immediately for me until I save, I wonder if maybe that's happening? Do you mind checking again and comparing [this version (instancing, webgl2)](https://editor.p5js.org/davepagurek/sketches/cWr5RmNAK) with [this version (using the CDN, no instancing, only webgl1)](https://editor.p5js.org/davepagurek/sketches/Zj1fJfUst)? If it's still slower with instancing, that's definitely a problem -- what browser/OS do you use?

> I also got an error when trying to use your version with webGL1, not sure if maybe that script got built before all of your changes were made, but just want to flag that.

Weird, I don't get that on Chrome/Firefox on my Mac, mind sending me the stacktrace of the error?