@davepagurek 2023-05-30
> I do worry a little about regressing performance on webGL1 though. Any way to prevent that?

This data format requires a tad more copying of values than before now that there are two position attributes instead of one. One option is to use a different line shader with different attributes in WebGL1, but I was hoping to avoid that because the maintenance cost goes up. I can see if there's anything I can do to speed up the copying in the WebGL1 case though. I can do a bit more debugging and see if something like using a fixed sized array upfront can help (since we can calculate the buffer sizes needed ahead of time before sorting the data into the non-instanced format)