@aferriss 2023-05-30
> Weird, I don't get that on Chrome/Firefox on my Mac, mind sending me the stacktrace of the error?

The error is `TypeError: r._makeStrokeBufferData is not a function
    at undefined:2:860144` and only occurs when drawing 3d geometry with calls like sphere or box. 
    
 I'm using chrome 113 on Mac OSX 13.13
 
I'm reliably seeing very low framerates when testing with just lines using your build, and around 20fps using the CDN build. I made a copy with some average framerate calculations here as well. 
https://editor.p5js.org/aferriss/sketches/HP0KpYo1J