@davepagurek 2023-05-30
Thanks! ok, found a fix for the crash for WebGL1 mode. Eventually I'll add some more testing for WebGL1 mode for retained and immediate mode, but it sounds like before getting to that there's some more fundamental performance things to work out.

Also thanks for the p5 editor link, I think the difference in my testing was on single shapes with lots of lines vs many shapes with single lines. It seems like when rendering shapes with a lot of data (the single large begin/endShape in my editor link), the bottleneck is in duplicating the data, which this change addresses, but when rendering a single line, the fact that there are more buffers involved means that the overhead of flattening multiple arrays and allocating more `Float32Array`s adds up. I'm going to spend some more time seeing if there's a way to keep the performance improvements for large shapes without compromising drawing many lines so much.