@davepagurek 2023-06-07
OK I'm back with some updated! I have a new version here for testing: https://editor.p5js.org/davepagurek/sketches/tcug_H1Z8

Updates I made to speed things up:
- **No longer checking if face culling is enabled when drawing strokes.** To make strokes work regardless of whether or not face culling is enabled, I updated the winding order of lines to be consistent with faces.
- **Using the same strides/divisors for all stroke buffers.** This seems to be a bit faster than changing it up between segments/caps/joins (maybe the WebGL driver internally does some deduplication of commands?)
- **Avoiding flattening data when putting it into RenderBuffers.** This is done by just writing them directly into an array rather than keeping around subarrays, and since the data is never read by other code (just sent right to the GPU), there is little cost in code readability.
- **Storing stroke data directly in `Float32Array`s.** I made a new `p5.DataVector` class similar to a C++ vector that doubles its capacity when data grows long. This way we avoid allocating new data every frame and having to GC it, which seems to be a big bottleneck.
  - As an aside, I also shrink the capacity down to be the nearest power of 2 before sending the data to the GPU, since it seems that even when you specify a count in `gl.bufferData`, it sends the whole `Flaot32Array`. Hopefully the power-of-2 sizes leave enough wiggle room to avoid too many resizes. It's great for drawing the same sort of shape many times in a row though!

Here's how it stands up against 1.6.0 in WebGL1 and 2 modes:

<table>
<tr><th rowspan=3>Test</th><th colspan=2>Previous</th><th colspan=4>Updated</th></tr>
<tr><th rowspan=2>Chrome</th><th rowspan=2>Firefox</th><td colspan=2>WebGL 1</td><td colspan=2>WebGL 2</td></tr>
<tr><th>Chrome</th><th>Firefox</th><th>Chrome</th><th>Firefox</th></tr>
<tr><th>1000 lines</th>
<td>60fps</td><td>16fps</td><td>(↓)28fps</td><td>(↑)24fps</td><td>(=)60fps</td><td>(↑)34fps</td></tr>
<tr><th>Large curve</th>
<td>23fps</td><td>10fps</td><td>(↓)12fps</td><td>(=)11fps</td><td>(↑)39ps</td><td>(↑)25fps</td></tr>
<tr><th>Retained+Immediate Shapes</th>
<td>60fps</td><td>60fps</td><td>(=)60fps</td><td>(=)60fps</td><td>(=)60fps</td><td>(=)60fps</td></tr>
</table>

Comparing against 1.6.0:
- In WebGL2 mode, everything is around 2x as fast, or capped out at the 60fps max
- In WebGL1 mode, Chrome is always 2x slower, while Firefox is around the same or marginally better

Personally, I think I'm OK with the WebGL1 tradeoffs, as I think its main use cases are:
- old browsers, e.g. Safari on my old Mac which I won't update the OS of, but it still has access to up-to-date Chrome and Firefox with WebGL2 support anyway
- shader compatibility with old shaders that used extensions (which could be used in WebGL2 by default but would require updating the GLSL syntax), but shader-based sketches seem not to rely much on geometry and lines

If you're OK with the performance tradeoffs for WebGL1, I can continue on and make another update with the extra comments/refactoring/tests we talked about earlier.