@aferriss 2023-06-08
Thanks for the updates @davepagurek ! I'll take some time to look through your new commits and test perf again. For my own clarity, in your table does the previous column indicate p5 perf prior to any of your changes, or do they represent the perf from your initial round of commits?  If it's the latter, including perf from prior to making any of these changes would be helpful to see. 

It looks like there are wins everywhere except for drawing in chrome using webGL1, and a more than 2x slow down still seems like a large regression to me. Do you have a sense of why it's such a substantial hit for that platform?

It'd be really great if we could test on some mobile devices, safari, and some windows machines as well if possible.